+++ docs/release-notes.md
@@ -8,6 +8,7 @@
 * Updated the Pintail library, greatly improving mod load times.
 * Fixed the game crashing when setting up a daily run with the Solo Run modifier with any modded characters.
 * The "current pile" indicator icons will no longer display if currently viewing a single pile.
+* The "sort cards in order" feature when browsing through cards in hand is now a proper sort mode, and it got moved from `Nickel.Bugfixes` to `Nickel.Essentials`.
 
 ### For developers:
 * Added `IModHelper.Content.Audio` which allows adding new sounds and music.
@@ -15,6 +16,7 @@
 * Added `PlayableCharacterConfigurationV2.SoloStarters`.
 * Added `(Non)PlayableCharacterConfigurationV2.Babble` which allows customizing the way the character talks in dialogue.
 * Nickel now validates whether the provided `ExeCardType` for a playable character uses the `Deck.colorless` deck.
+* Added a `Nickel.Essentials` API to control whether the "sort cards in order" sort mode is enabled for the given `CardBrowse` instance.
 
 ## 1.7.0
 Released 17 December 2024.